WIP: refactoring of derived_model; split into single function files
