+++ resqpy/derived_model/__init__.py
@@ -1,15 +1,44 @@
-__all__ = ['add_one_blocked_well_property',
-           'add_one_grid_property_array',
-           'add_wells_from_ascii_file',
-           'add_zone_by_layer_property',
-           'interpolated_grid',
-           'zonal_grid',
-           'zone_layer_ranges_from_array']
+__all__ = ['dm_add_edges_per_column_property_array'.
+           'dm_add_faults',
+           'dm_add_one_blocked_well_property',
+           'dm_add_one_grid_property_array',
+           'dm_add_single_cell_grid',
+           'dm_add_wells_from_ascii_file',
+           'dm_add_zone_by_layer_property',
+           'dm_coarsened_grid',
+           'dm_common',
+           'dm_copy_grid',
+           'dm_drape_to_surface',
+           'dm_extract_box_for_well',
+           'dm_extract_box',
+           'dm_fault_throw_scaling',
+           'dm_gather_ensemble',
+           'dm_interpolated_grid',
+           'dm_local_depth_adjustment',
+           'dm_refined_grid',
+           'dm_tilted_grid',
+           'dm_unsplit_grid',
+           'dm_zonal_grid',
+           'dm_zone_layer_ranges_from_array']
 
-from .add_one_grid_property_array import add_one_grid_property_array
-from .add_one_blocked_well_property import add_one_blocked_well_property
-from .add_wells_from_ascii_file import add_wells_from_ascii_file
-from .add_zone_by_layer_property import add_zone_by_layer_property
-from .interpolated_grid import interpolated_grid
-from .zonal_grid import zonal_grid, single_layer_grid
-from .zone_layer_ranges_from_array import zone_layer_ranges_from_array
+from .dm_add_edges_per_column_property_array import add_edges_per_column_property_array
+from .dm_add_faults import add_faults
+from .dm_add_one_blocked_well_property import add_one_blocked_well_property
+from .dm_add_one_grid_property_array import add_one_grid_property_array
+from .dm_add_single_cell_grid import add_single_cell_grid
+from .dm_add_wells_from_ascii_file import add_wells_from_ascii_file
+from .dm_add_zone_by_layer_property import add_zone_by_layer_property
+from .dm_coarsened_grid import coarsened_grid
+from .dm_copy_grid import copy_grid
+from .dm_drape_to_surface import drape_to_surface
+from .dm_extract_box_for_well import extract_box_for_well
+from .dm_extract_box import extract_box
+from .dm_fault_throw_scaling import fault_throw_scaling, global_fault_throw_scaling
+from .dm_gather_ensemble import gather_ensemble
+from .dm_interpolated_grid import interpolated_grid
+from .dm_local_depth_adjustment import local_depth_adjustment
+from .dm_refined_grid import refined_grid
+from .dm_tilted_grid import tilted_grid
+from .dm_unsplit_grid import unsplit_grid
+from .dm_zonal_grid import zonal_grid, single_layer_grid
+from .dm_zone_layer_ranges_from_array import zone_layer_ranges_from_array
